Create repro cases for FBO clear bug
Color clears typically enjoy special handling in drivers, especially on
tile-based renderers, where fullscreen clears are typically just state
that gets processed when tile is initialized. That requires further
special handling when the clear is the only command issued to a render
target.

Sure enough there appears to be some devices where following doesn't
work as expected:

Repeat N times:
  Clear FBO A with some iteration-specific color
  Bind FBO B
  Sample from FBO A

The result is that sampling might not return color of that iteration,
but instead for example the claer color of first, or previous iteration.

These tests attempt to reproduce that implementation bug.

Bug: 18169662
Change-Id: I869138c5108d66fdf0cf497f730b2b3518d1686d