Relax fma() verification in builtin_functions.common.fma
GL_EXT_gpu_shader5 allows fma() to be implemented either as a single, or
two correctly-rounded operations. Current tests assumed fma() for lowp
and mediump to be implemented as a single operation as no rounding was
allowed mid-operation.

This change relaxes fma() tests to allow for a*b+c expansion.

Bug: 19431850
Change-Id: I46eea74bf3b95bfdbc510e13ef7315854be623ee