Fix EGL multithread single window tests.
Avoid trying to destroy windows when current thread has no windows in single
window cases. Tests always create new resources, if there are too few, and
destroy resources when there are too many. When number of resources is somewhere
in between test chooses randomly what to do next. In a single window case the
current thread may not be able to create a new window and thus may not have any
windows. When there are no windows the random operation must not attempt to
destroy a window.

Bug: 20958064
Change-Id: I5366233c141278a994768110b399afa79f6ed2e5