Remove non-trivially modified random shader cases from mustpass.
Random shader generation erroneusly caused int overflows in value range
calculations. This is undefined behavior and it was fixed in commit
2106d67 (Change-Id: I810be9c747504df671629201ba56b5b8ed2307d7) but
since the existing tests were generated using the broken code, the fix
changed slightly the generation logic and thus the existing tests. This
CL removes all tests from mustpass for which the generation results in
significantly different shaders. Trivially changed shaders are left
unremoved as the feature coverage of those tests is expected to remain
unchanged.

Bug: 21851700
Change-Id: Ic0930602fc17c1851b9ffa694c04a884294f438b