Don't mangle safe bit patterns in copy_image sRGB RB cases.
Clearing a sRGB framebuffer takes in the clear color in linear space and
then transforms it to sRGB space. This process destroys the carefully
selected clear color bit pattern. Avoid bit mangling by doing first an
inverse tranformation to the supplied clear color which restores the bit
pattern when the color is coverted to sRGB internal format.

Bug: 20698928
Bug: 21951463
Change-Id: I2155e8a977cc524401b5f98c7dde4502f62e35a7