buffer_view.create: Only create views after binding memory
The only time the Vulkan API allows you to create a view after binding
memory is if you are using sparse memory.  However, sparse memory requires
a different binding paradigm.  Since we don't actually support sparse
binding in this test, we're probably better off just not claiming to for
now and fixing the binding problem.

This fixes #84