api.binding_model: Wait for idle before throwing NotSupportedError
If we don't, then the GPU could be busy and some of the objects destroyed
by the exception propagation could be in-use.  This is something of a
stop-gap measure; the problem of exceptions and the GPU being busy is
something that needs to be solved in general.  However, it does fix GPU
hangs on some drivers.

This is related to #20.