SSBOLayoutCase: Write inout descriptors with a single array
The Vulkan spec allows descriptor writes that go over the bounds of the
current binding to spill into the next as long as the bindings have the
same type.  Using this allows us to make the update a bit simpler, and will
keep it correct when we start using more complex layouts.