Don't invoke implementation-defined behavior in drawRangeElements tests
GLES 3.0.4 section 2.9.3 page 31 about drawRangeElements: "It is an
error for index values to lie outside the range [start, end], but
implementations are not required to check for this."

ANGLE currently does check for this and generates an error in this
case as allowed by the spec. Initialize memory passed as indices to
drawRangeElements, so that dEQP doesn't invoke this behavior when it's
expecting the call to generate NO_ERROR. In the cases where dEQP does
expect drawRangeElements to generate an error, the memory passed as
indices is initialized to 0xffffffff.

Change-Id: I495a55d69d9b1fe12c335fca66fd1e1d23c9a232