Install APK in the deqp runner to reduce installs
Currently the deqp APK is installed both by preparer in the config and
the runner class. Using preparer only has the drawback that the preparer
is run for each config independent of filtering results. Doing the
installation at the runner enables the runner skip the lengthy
installation if no tests are executed for config.

Reduces the execution time for deqp tests with null filter from minutes
to 7 seconds.

Tested on Angler on nyc-dev

Bug: 27810566
Change-Id: Ia9cdc88bbc9bdb737d5576319b954cd7fbbc902b