Fix required sample count tests
According to,

   Table 11.1. Image and image view parameter compatibility requirements

Only non-cube images with type VK_IMAGE_TYPE_2D can have sample counts
greater than one.

In addition, the following properties should hold for multisampled
images: the formats must support being used in a color or depth/stencil
attachment, or in a storage image when shaderStorageImageMultisample is
supported. Images not fitting the above descriptions may have a sample
count of one or zero (in error conditions). Align tests for sample
counts to this behavior.