GL_ARB_gpu_shader_fp64: Forces UBOs to use the std140 layout
The max_uniform_components test for the FP64 extension was asuming
UBOs were stored in a tightly packed fashion. The way this is done is
implementation dependent; for example, mesa uses std140. The only way
of making any assumption on how an implementation is packing the data
for UBOs is, actually, by specifying the std140 layout.

Unfortunately, this also involves that the available used space for a
single UBO may be exhausted much before we get to use the maximum
uniform components. Hence, this patch also adds a check for
GL_MAX_UNIFORM_BLOCK_SIZE and limits the generated test to use the
amount of uniforms that fit in whichever limit gets exhausted first.

Fixes the following test (from fail to pass):
     GL44-CTS.gpu_shader_fp64.max_uniform_components

Change-Id: I734b090dbded4899e4db48c969754ab0a67c5245