Test descriptor_update_template, push_descriptor
Extended the existing binding_model tests with 3 new variants,
'with_template', 'with_push_template' and 'with_push'.

The least invasive method to support push descriptor updates without
template was to persist the DescriptorSetUpdateBuilder object in the
test instance classes, so that it can be referenced when building the
command buffer.

Change-Id: I731fa41c5ff99ab1d390d77d7aabb4424e091e35