Fix uninitialized char arrays in ProgramInterfaceQueryTests.
Replicated from OpenGL/CTS Merge Request 432.

This patch initializes all of the char arrays in 
the ProgramInterfaceQueryTests in the OpenGL and OpenGL ES versions of 
conformance. Since they are char arrays
we should initialize them to the null terminator.

The motivation here is to prevent the verification script from 
encountering unparsable tokens in the qpa result files from conformance.

This patch fixesOpengGL OSS-CTS issue 4.

Change-Id: Ia95742db876520d911b91039402dee46bbc6ebfe