Move MatrixLoadFlags from vktTestCase.hpp
vktTestCase.hpp is wildly inappropriate place for such data type. Both
UBO and SSBO tests use it currently and it is now duplicated to both of
them. UBO and SSBO tests do contain a quite bit of duplication already
(describing blocks, layout computation, etc.) and dedup should be done
as a part of bigger refactoring.

Change-Id: I5d4b8a947522192ba0dd4e807e04c940c1770eef