direct-state-access/texture-tests: Use a type compatible with the format
One of the test sub-cases in ImageQueryErrorsTest family of tests is using
an incompatible format/type, when it is just checking for format and
internal-format compatibility. This is adding another failure point to the
test, making it dependent of the order of validation of the arguments in
the OpenGL driver.

Specifically, the sub-case that checks that "format is DEPTH_STENCIL and the
base internal format is not DEPTH_STENCIL" is passing GL_DEPTH_STENCIL with
type GL_UNSIGNED_INT, which would fail with INVALID_ENUM:

From OpenGL 4.5, section '18.2 Reading Pixels', page 506 of the PDF states:

    "An INVALID_ENUM error is generated if format is DEPTH_STENCIL and
     type is not UNSIGNED_INT_24_8 or FLOAT_32_UNSIGNED_INT_24_8_REV."

This is for ReadPixels, but the same pixel storage modes apply for
GetTextureImage:

From OpenGL 4.5, section '8.11 Texture Queries', page 234 of the PDF states:

    "Pixel storage modes that are applicable to ReadPixels are applied, as
     described in table 18.1 and section 18.2.9."

This patch changes the type to GL_UNSIGNED_INT_24_8, which is compatible with
the format, and will allow the subcase to unambiguosly fail if format and
internal-format are incompatible.

Change-Id: I979e007aef6e09b1707605bf4b94aebaa2d21b1c