es31: complete check for rgba32f internalformat
Right now several tests check for the extension
EXT_color_texture_buffer to know if that internalformat is
supported. But at this moment that internalformat is also supported on
core profiles, or if ARB_color_texture_buffer is available.

Note that it is not possible to just add a check for
ARB_color_texture_buffer. On core profiles, most of the functionality
got included on the spec, but deprecating some of the enums. As an
outcome, some drivers decide to not expose the extension, even if the
internalformat and other functionality is exposed.

For more info, take a look to the following commits on mesa:
3264c3e99700389f0a3958db7c9c19673107d67a
bd850cb4f2c77e2eb6716c865c40b9976633fc23

This commit gets 72 tests going from skipped to pass on
Intel Skylake and Broadwell:
GL45-CTS.sample_variables.mask.rgba32f.*
GL45-CTS.shader_multisample_interpolation.render.*.rgba32f.*

Also move es31cSampleVariableTests and es31cShaderMultisampleInterpolationTests
to common.

Tests defined on those files are used on both GLES and GL test case lists,
so it makes sense to move them to the common module.

Change-Id: I0ed09e9897f9f2a1b4efd1a40c22da00ec70bfdc