ImageLoadStoreMultipleUniformsTest: fix reversed bit-field overflow check
This class has a templated Mask struct used to store and access
bit-fields. In includes a specialization that is used if the offset is
bigger that the bitfield, that raises assert.

Without this patch, the check is reversed, so it fails on cases where
the offset is correct.

This patch changes the condition. An alternative would be change the
value of the specialization with asserts. I found this option more
natural.

This fixes the following test:
GL45-CTS.shader_image_load_store.multiple-uniforms

on the Mesa driver for Intel Haswell hw. It makes it goes from
crash to fail on Broadwell/Skylake.

Change-Id: I7341a542ad925e55ab8a6c95373567b4ea58434f