opengl/khr_debug: Push/pop the custom render context
KHR_Debug tests create their own render context. However, the
framework's test iterator handles basic operations on the render context
between test executions (e.g, swapping buffers), but it always assumes
that the default render context is the one being used. This causes
conflicts on KHR_debug tests, where the postIterate() method throws a
ResourceError, aborting the test execution, and breaking the logging
system.

This patch changes all KHR_debug tests to first save the original render
context, then sets the created render context as that of the base
context to give the framework a good, active render context; and finally
sets back the riginal render context when the test finishes.

If the test aborts in the middle, the destructor of the test (in TestBase)
will still try to restore the original context.

To allow the above, a setRenderContext() method was added to the test
context. So we need to have a mechanism to temporarily set the render
context of the default test context during the execution of these tests.

Fixes all tests in:
* GLXX-CTS.khr_debug.*

Fixes #28

Change-Id: If2c3eea5cb0882c11d2baf8334ed9ee6208b6c65