Fix assert for combined depth/stencil formats
This was using tcu::isCombinedDepthStencilType(format.type) for this,
but this only checks the type and for 24bit depth formats with a
x8 d24 setup it will incorrectly identify them as combined
depth/stencil even though the underlying format is depth-only. Fix
this by testing if the format has depth and stencil components
instead.

Prevents crashes in tests with the following signature (144 tests):
dEQP-VK.texture.shadow.*x8_d24*

Fixes #563

Change-Id: Ie43e140a0461de49c02e0d97eeed38fa8ef24040