geometry_shader/api: Fix linking error
The tests redeclared the built-in 'gl_PerVertex' interface block in the
Vertex shader when run in a GL context. However, they did not redeclare
'in gl_PerVertex gl_in[]' and they used it, leading to an undesirable
linking error which is not what the test wished to check.

The patch adds the 'in gl_PerVertex gl_in[]' redeclaration to the
Geometry shader when the tests are run under GL.

From the GLSL 4.5 specification, "7.1 Built-In Language Variables"
(page 130):

"If multiple shaders using members of a built-in block belonging to
the same interface are linked together in the same program, they
must all redeclare the built-in block in the same way, as described
in section 4.3.9 “Interface Blocks” for interface-block matching, or
a link-time error will result. It will also be a link-time error if
some shaders in a program redeclare a specific built-in interface
block while another shader in that program does not redeclare that
interface block yet still uses a member of that interface block. If
a built-in block interface is formed across shaders in different
programs, the shaders must all redeclare the built-in block in the
same way (as described for a single program), or the values passed
along the interface are undefined."

VK-GL-CTS Issue: 130

Components: OpenGL

Affects:

GL44-CTS.geometry_shader.api.insufficient_emitted_vertices
GL44-CTS.geometry_shader.api.incompatible_draw_call_mode
GL45-CTS.geometry_shader.api.insufficient_emitted_vertices
GL45-CTS.geometry_shader.api.incompatible_draw_call_mode

Change-Id: I7d5f54b287b778e46376ef6ac9effcd90778705c