Fix crashes in dEQP-VK.glsl.matrix.inverse.*
This is a regression caused by altering the value of the GRID_SIZE
in commit 2ba66fa2, which results in the inverse() tests generating
input matrices which determinant is zero, and thus, cannot be
inverted, which makes the tests hit an assertion.

This patch adds a mechanism for shader tests to specify a grid size
different from the default. This allows shader testst that are
sensitive to this value, such as matrix inverse tests, to define
their own requirements.

Fixes VK-GL-CTS isssue: 66
Components: Vulkan

Affects:
dEQP-VK.glsl.matrix.inverse.*

Change-Id: I31a34c2d156ee5b98d2de1b10ac4cec1a90d480e