Removed GL45-CTS.draw_elements_base_vertex_tests.invalid_mapped_bos
Removed GL45-CTS.draw_elements_base_vertex_tests.invalid_mapped_bos
The test is made to fail if a draw call on a mapped buffer is made.  The
specs for OpenGL and ES in sections 6.3.2 stated the draw call may fail
but do not suggest a mandatory INVALID_OPERATION to be thrown.