Write gl_PointSize value when it is undefined
gl_PointSize is undefined unless expicitly written to. The only
exception to this rule is the case when
shaderTessellationAndGeometryPointSize is disabled, in which case it
defaults to 1.0 for these stages.

Since we don't have access to the physical device features at shader
compile time, implement two variants of the tessellation evaluation
shader when point mode is enabled, one that writes gl_PointSize and
one that doesn't and decide the one we need at run time based on
whether the feature is enabled or not.

Components: Vulkan
Fixes VK-GL-CTS issue: 159

Affects:
dEQP-VK.tessellation.primitive_discard.*point_mode

Change-Id: Idadbc70a704f00ed601ff65afeb2a3edfbeb3f19