Fix [Named]FramebufferTextureLayer tests
The test was failing because it was reusing the texture ID used while
testing FramebufferTexture. But that is a GL_TEXTURE_2D, that is a
wrong target for [Named]FramebufferTextureLayer. So for the other
subtests, that were checking things like invalid attachment, etc,
there was two possible errors, and mesa driver raised first the one
related to a wrong target.

This patch reuses a texture ID with a valid target, being also used to
test [Named]FramebufferTextureLayer

For what it is worth, [Named]FramebufferTextureLayer with a wrong
target is also tested.

Affects:
GL45-CTS.direct_state_access.framebuffers_texture_attachment_errors

Components: OpenGL

VK-GL-CTS issue: 173

Change-Id: I388e1593e67d8708433d0f4763c716a4f3f03067