Explicitly use std140 layout for uniform blocks
The exact memory layout of an array of uints in a uniform block is
implementation-defined, unless a layout is explictly specified.
So the UniformBufferStorageTestCase would previously fail on an
implementation that uses std140 by default.

Use std140 explicitly -- which implies changing the stride of the
array to 16 bytes -- since std140 is the only layout supported for
uniform blocks in core OpenGL.

Modules: OpenGL
Affects:
GL45-CTS.sparse_buffer_tests.BufferStorageTest

VK-GL-CTS issue: 179

Change-Id: Ib5a996aa55a4c589625c575a1b4a3d252ca8e30c