Fix post-depth coverage shader test case
The tests uses 4 different versions of GLSL shader code for
the fragment shader that are supposed to be assigned to the
variables m_fragShader{1..4} that are then used in the 4
test cases implemented. However, the last shader code is not
assigned to m_fragShader4, instead it immediately overwrites
m_fragShader3. The test would then fail as it attempts to build
a shader program with an empty string for the fragment shader
when it uses m_fragShader4.

Components: OpenGL
VK-GL-CTS issue: 181

Affects:
GL45-CTS.post_depth_coverage_tests.PostDepthShader

Change-Id: I49474f04668de2037b4afc5570082bd4110f4fab