DirectStateAccessTexturesTests: fix for textures_subimage_errors
One of the subtests for compressed textures 1D, 2D and 3D are
checking the following error from spec:

  "An INVALID_OPERATION error is generated if format does not match
   the internal format of the texture image being modified, since
   these commands do not provide for image format conversion."

In order to do that it used a not-matching format. But the problem is
that it didn't change the image size, so a different error (checked
later on the test) was generated.

This commit solves the problem by asking to the driver the size for
the not-matching format, in the same way that the size is asked for
the reference formats.

Affects:
GL45-CTS.direct_state_access.textures_subimage_errors

Components: OpenGL

VK-GL-CTS issue: 188

Change-Id: If0f17437dd8767d22389770068cde55dd36e2602