CullDistanceTests/coverage: Fix a couple issues in the shaders
There are two specific problems in the shaders of this test, both
preventing the fragment shader from receiving any fragment (no geometry
is renderer), and thus the gl_maxCullDistances value read back from the
texture is always zero.

The first issue is in the geometry shader. The 4 emitted vertices have
a 0.0 normal, the quad is not visible.

The second issue is in the tessellation evaluation shader, where the
value of gl_Position is not explicitly set, so the point is rendered
at an undefined location.

Components: OpenGL

Affects:
GL45-CTS.cull_distance.coverage

VL-GL-CTS issue: 197

Change-Id: I145894192fe9504808666e2c2fcd6cf53dbb03df