Further relax line verification in primitive bbox tests
The original relaxation patch
ba45591407d0207e8ba0634ea5636fd721ed781f
allowed lines to be 1 pixel thinner or thicker compared
to the expected width. Unfortunately it missed the case
of a line with the width of 1 pixel that has extra verification
in the framework. This change covers this gap.

Google bug: 35987764

Change-Id: I193d76ae13738702c9d7361d1957f15ea7600c4f