Add GL/ES VERSION defines
The dEQP does not use the standard GL or GLES headers, but instead
provides it's own type and enum definitions via glw.h.
However this was missing the GL_VERSION_* and GL_ES_VERSION_* defines
that are defined in the standard headers when the types are defined.
This change adds a new inlined include file 'glwVersions.inl' that
defines the GL_*VERSION* for the API versions that are currently
enabled.

VK-GL-CTS issue: 196
Components: OpenGL, Framework

Change-Id: Ie3f9a8e3958aee2cfdee4a4bc6de392bfeb13f1d