Check DynamicIndexing in pipeline.image count tests
The pipeline.image.sampling_type*count_* tests where the count of images
is greater than 1 use an array of images/samplers which requires the use
of shaderSampledImageArrayDynamicIndexing, which the test does not
currently check for.

Adds a check and throws NotSupportedError if the feature is not present.

Affects:

dEQP-VK.pipeline.image.sampling_type*count_*

Components: Vulkan

VK-GL-CTS issue: 296

Change-Id: I7b4244bcf4371e41c2121812362db166dc1d4c16