Add support for metamorphic testing
Add a new test case which compare the renderings of two different yet
semantically equivalent fragment shaders. A couple of shadersets are
also added as associated data, each contains one recipient fragment
shader which is the reference, and a few variants which are semantically
equivalent to the recipient. Each variant included here has proven to
trigger some issue on an Android device, see comment in the variant
source code.

Besides comparing the renderings of both recipient and variant shaders,
the test case also checks that each shader leads to a deterministic
rendering, by performing rendering twice and doing a pixel-per-pixel
comparison on the resulting image. When comparing the recipient and
variant renderings, a slight difference is tolerated to cope with
variations due to floating point arithmetics. Moreover, a sanity check
function may be used in order to verify that the recipient image is
relevant, and thus avoid a e.g. full black image rendering to pass the
test.

Change-Id: Id6cbbe7459864c6fb3acf72faa393ee0f4d8315d