Remove wrong image barrier from multisample_shader_builtin tests
Spec says:
"If vkCmdPipelineBarrier is called within a render pass instance,
the oldLayout and newLayout members of an element of
pImageMemoryBarriers must be equal"

The `multisample_shader_builtin` tries to use barrier for changing
the layout of the image used as render target and input attachment.
The barrier is invalid and, given the existing subpass dependencies,
not needed at all.

VK-GL-CTS Issue: 386
Component: Vulkan
Affects: dEQP-VK.pipeline.multisample_shader_builtin.*

Change-Id: I23bf1de5d39a5d786a6bf863a429f8d7e97f495d