Remove copy_image special path for rgb9e5/rg11b10
Copy_image is similar to a memcpy. When analyzing copy_image result,
the test have a special path for rgb9e5 and rg11b10.
Instead of using a memcmp, it converts each component to float and
then compare values.

This could be problematic, when data from other formats is
reinterpreted as rgb9e5/rg11b10, the result can be INF or NAN.

Doing a float equal comparison on 2 NAN values will return false,
so we can have 2 identical 32-bits values that are considered
different by the test because they are reinterpreted as float.

This change remove the special path for rgb9e5 and rg11b10.
I don't think it is required since copy_image should be like a
memcpy, a memcmp should be enough to verify the result

Affects:

KHR-GL45.copy_image.functional

Components: OpenGL

VK-GL-CTS issue: 404

Change-Id: Ifc7f7b4c8d62692e9364510ebc4e4162468e9568