Skip height validation on TEXTURE_1D_ARRAY
In verifyTexStorageInvalidValueErrors the test validate that a
GL_INVALID_VALUE error is returned by the GL when the height
parameters of the texStorage call is invalid.

The height is ignored by the texStorage call for 1d array, so the
error is never generated.

We should skip this validation for 1d array

Affects:

KHR-GL45.sparse_texture_tests.*

Components: OpenGL

VK-GL-CTS issue: 437

Change-Id: Ibe93db467deb560f5394b4b658dfa3f32a953709