Fix incorrect buffer size expectation in DSA XFB buffer test
The test creates two buffers, both with a size 's_bo_size' in
BuffersTest::prepareTestSetup(), then binds the first buffer
to binding point 0 and each half of the other buffer to
binding points 1 and 2 respectively.

Later, the test queries buffer sizes for each binding and
expects the size of the buffer in the first binding to be 0
when it should be 's_bo_size'.

Components: OpenGL
VK-GL-CTS issue: 502

Affects:
KHR-GL45.direct_state_access.xfb_buffers

Change-Id: I394622bb7bba076c68c5cca968c6b1cbf34e2ea9