Bump GLSL version for gl_PerVertex validation test
While gl_PerVertex was introduced into desktop GL in version 3.2, it
wasn't legal to redeclare it until after ARB_separate_shader_objects,
which became part of core in GL 4.1. This change bumps shader versions
up to GL 4.1 so we don't have to check for ARB_separate_shader_objects
all over the place.

Also add an error message in the event a subtest fails, so that it's
easier to identify which subtest(s) failed.

Components: OpenGL
VK-GL-CTS issue 505

Affects:

* KHR-GL41.CommonBugs.CommonBug_PerVertexValidation
* KHR-GL42.CommonBugs.CommonBug_PerVertexValidation
* KHR-GL43.CommonBugs.CommonBug_PerVertexValidation
* KHR-GL44.CommonBugs.CommonBug_PerVertexValidation
* KHR-GL45.CommonBugs.CommonBug_PerVertexValidation

Change-Id: I81405952e186377cb617e2d77063d09a0f5ce8fc