Rewrite DSA texture buffer tests to signal unsupported configurations
The current implementation has a single test instance that
loops over a collection of types and sizes and produces a
single pass/fail result. Because some of the framebuffer
configurations produced may not be supported by the driver,
the test skips silently such configurations, which is not
desirable.

This patch rewrites the implementation so that we have a
separate test instance for each configuration instead. This
allows us to signal independent pass/fail/unsupported results
for each one.

Components: OpenGL

Affects:
KHR-GL45.direct_state_access.textures_buffer

Change-Id: Id160dd91bed4e5e2a66a6e579e9e4893b5ea956a